String Word Count
String word count could also be used to break a string into an array with each word being assigned to an index.  This is been coded in but commented out because the explode command seems to do the same job but without the need to specify characters to be included.